auth service

+++ src/users/user.service.ts
@@ -2,9 +2,11 @@ import {
   BadRequestException,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { generateAuthToken } from 'src/helpers/generateAuthToken';
+import { JwtService } from '@nestjs/jwt';
+// import { generateAuthToken } from 'src/helpers/generateAuthToken';
 import { hashPassword } from 'src/helpers/hashPassword';
 import { Order, User } from 'src/typeorm';
 import { Repository } from 'typeorm';
@@ -15,11 +17,11 @@ export class UsersService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
+    private jwtService: JwtService,
   ) {}
   async createUser(createUserDto: CreateUserDto): Promise<string> {
     const { email, password } = createUserDto;
     const existsUser = await this.userRepository.findOneBy({ email });
-    console.log(existsUser);
     if (existsUser) {
       throw new BadRequestException('Email already exists');
     }
@@ -30,22 +32,29 @@ export class UsersService {
     return 'User successfully created';
   }
 
-  getUsers() {
-    return this.userRepository.find({
-      select: ['id', 'name', 'email', 'role', 'order'],
-      relations: ['order'],
-    });
+  getUsers(req): Promise<Partial<User[]>> {
+    const { role } = req.user;
+    if (role == 'admin') {
+      return this.userRepository.find({
+        select: ['id', 'name', 'email', 'role'],
+      });
+    }
+    throw new UnauthorizedException();
   }
 
-  async getSingleUser(userId: string) {
-    const user = await this.findUser(userId);
-    return {
-      id: user.id,
-      name: user.name,
-      email: user.email,
-      role: user.role,
-      order: user.order,
-    };
+  async getSingleUser(userId: string, req): Promise<Partial<User>> {
+    const { role } = req.user;
+    if (role == 'admin') {
+      const user = await this.findUser(userId);
+      return {
+        id: user.id,
+        name: user.name,
+        email: user.email,
+        role: user.role,
+        order: user.order,
+      };
+    }
+    throw new UnauthorizedException();
   }
 
   async login(email: string, password: string) {
@@ -59,25 +68,28 @@ export class UsersService {
     if (user.password !== hashedPassword) {
       throw new BadRequestException('Invalid credentials');
     }
-    const token = await generateAuthToken(user.id);
+    const payload = { sub: user.id, name: user.name, role: user.role };
+    const token = await this.jwtService.signAsync(payload);
     return {
       token,
       user: {
         id: user.id,
         name: user.name,
-        email: user.email,
-        role: user.role,
       },
     };
   }
 
-  async deleteUser(userId: string) {
-    const user = await this.findUser(userId);
-    this.userRepository.delete(user);
-    return 'User deleted';
+  async deleteUser(userId: string, req): Promise<string> {
+    const { role } = req.user;
+    if (role == 'admin') {
+      const user = await this.findUser(userId);
+      this.userRepository.delete(user);
+      return 'User deleted';
+    }
+    throw new UnauthorizedException();
   }
 
-  async findUser(id: string) {
+  async findUser(id: string): Promise<User> {
     const user = await this.userRepository.findOne({
       where: { id },
       select: ['id', 'name', 'email', 'role', 'order'],
